Revert "Feature/add players team relation and seed"
