--- models/Team.js
@@ -1,25 +1,0 @@
-const mongoose = require("mongoose");
-const Schema = mongoose.Schema;
-
-const playerSchema = new Schema(
-  {
-    name: {
-      type: String,
-      required: true,
-    },
-    foundation: {
-      type: Number,
-      required: true,
-    },
-    city: {
-      type: String,
-      required: true,
-    },
-  },
-  {
-    timestamps: true,
-  }
-);
-
-const Artist = mongoose.model("Team", artistSchema, "teams");
-module.exports = { Team };
